+++ integrations/langfuse/tests/test_tracer.py
@@ -9,7 +9,7 @@ from unittest.mock import MagicMock, Mock, patch
 
 import pytest
 from haystack.dataclasses import ChatMessage
-from haystack_integrations.tracing.langfuse.tracer import LangfuseTracer, SpanContext, DefaultSpanHandler
+from haystack_integrations.tracing.langfuse.tracer import LangfuseTracer, LangfuseSpan, SpanContext, DefaultSpanHandler
 
 
 class MockSpan:
@@ -43,6 +43,149 @@ class MockTracer:
 
     def flush(self):
         pass
+
+
+class TestLangfuseSpan:
+
+    #  LangfuseSpan can be initialized with a span object
+    def test_initialized_with_span_object(self):
+        mock_span = Mock()
+        span = LangfuseSpan(mock_span)
+        assert span.raw_span() == mock_span
+
+    #  set_tag method can update metadata of the span object
+    def test_set_tag_updates_metadata(self):
+        mock_span = Mock()
+        span = LangfuseSpan(mock_span)
+
+        span.set_tag("key", "value")
+        mock_span.update.assert_called_once_with(metadata={"key": "value"})
+        assert span._data["key"] == "value"
+
+    #  set_content_tag method can update input and output of the span object
+    def test_set_content_tag_updates_input_and_output(self):
+        mock_span = Mock()
+
+        span = LangfuseSpan(mock_span)
+        span.set_content_tag("input_key", "input_value")
+        assert span._data["input_key"] == "input_value"
+
+        mock_span.reset_mock()
+        span.set_content_tag("output_key", "output_value")
+        assert span._data["output_key"] == "output_value"
+
+    # set_content_tag method can update input and output of the span object with messages/replies
+    def test_set_content_tag_updates_input_and_output_with_messages(self):
+        mock_span = Mock()
+
+        # test message input
+        span = LangfuseSpan(mock_span)
+        span.set_content_tag("key.input", {"messages": [ChatMessage.from_user("message")]})
+        assert mock_span.update.call_count == 1
+        # check we converted ChatMessage to OpenAI format
+        assert mock_span.update.call_args_list[0][1] == {"input": [{"role": "user", "content": "message"}]}
+        assert span._data["key.input"] == {"messages": [ChatMessage.from_user("message")]}
+
+        # test replies ChatMessage list
+        mock_span.reset_mock()
+        span.set_content_tag("key.output", {"replies": [ChatMessage.from_system("reply")]})
+        assert mock_span.update.call_count == 1
+        # check we converted ChatMessage to OpenAI format
+        assert mock_span.update.call_args_list[0][1] == {"output": [{"role": "system", "content": "reply"}]}
+        assert span._data["key.output"] == {"replies": [ChatMessage.from_system("reply")]}
+
+        # test replies string list
+        mock_span.reset_mock()
+        span.set_content_tag("key.output", {"replies": ["reply1", "reply2"]})
+        assert mock_span.update.call_count == 1
+        # check we handle properly string list replies
+        assert mock_span.update.call_args_list[0][1] == {"output": ["reply1", "reply2"]}
+        assert span._data["key.output"] == {"replies": ["reply1", "reply2"]}
+
+
+class TestSpanContext:
+    def test_post_init(self):
+        with pytest.raises(ValueError):
+            SpanContext(name=None, operation_name="operation_name", component_type=None, tags={}, parent_span=None)
+        with pytest.raises(ValueError):
+            SpanContext(name="name", operation_name=None, component_type=None, tags={}, parent_span=None)
+        with pytest.raises(ValueError):
+            SpanContext(
+                name="name",
+                operation_name="operation_name",
+                component_type=None,
+                tags={},
+                parent_span=None,
+                trace_name=None,
+            )
+
+
+class TestDefaultSpanHandler:
+    def test_handle_generator(self):
+        mock_span = Mock()
+        mock_span.raw_span.return_value = mock_span
+        mock_span.get_data.return_value = {
+            "haystack.component.type": "OpenAIGenerator",
+            "haystack.component.output": {"replies": ["This the LLM's response"], "meta": [{"model": "test_model"}]},
+        }
+
+        handler = DefaultSpanHandler()
+        handler.handle(mock_span, component_type="OpenAIGenerator")
+
+        assert mock_span.update.call_count == 1
+        assert mock_span.update.call_args_list[0][1] == {"usage": None, "model": "test_model"}
+
+    def test_handle_chat_generator(self):
+        mock_span = Mock()
+        mock_span.raw_span.return_value = mock_span
+        mock_span.get_data.return_value = {
+            "haystack.component.type": "OpenAIChatGenerator",
+            "haystack.component.output": {
+                "replies": [
+                    ChatMessage.from_assistant(
+                        "This the LLM's response",
+                        meta={"model": "test_model", "completion_start_time": "2021-07-27T16:02:08.012345"},
+                    )
+                ]
+            },
+        }
+
+        handler = DefaultSpanHandler()
+        handler.handle(mock_span, component_type="OpenAIChatGenerator")
+
+        assert mock_span.update.call_count == 1
+        assert mock_span.update.call_args_list[0][1] == {
+            "usage": None,
+            "model": "test_model",
+            "completion_start_time": datetime.datetime(2021, 7, 27, 16, 2, 8, 12345),
+        }
+
+    def test_handle_bad_completion_start_time(self, caplog):
+        mock_span = Mock()
+        mock_span.raw_span.return_value = mock_span
+        mock_span.get_data.return_value = {
+            "haystack.component.type": "OpenAIChatGenerator",
+            "haystack.component.output": {
+                "replies": [
+                    ChatMessage.from_assistant(
+                        "This the LLM's response",
+                        meta={"model": "test_model", "completion_start_time": "2021-07-32"},
+                    )
+                ]
+            },
+        }
+
+        handler = DefaultSpanHandler()
+        with caplog.at_level(logging.ERROR):
+            handler.handle(mock_span, component_type="OpenAIChatGenerator")
+            assert "Failed to parse completion_start_time" in caplog.text
+
+        assert mock_span.update.call_count == 1
+        assert mock_span.update.call_args_list[0][1] == {
+            "usage": None,
+            "model": "test_model",
+            "completion_start_time": None,
+        }
 
 
 class TestLangfuseTracer:
@@ -170,88 +313,3 @@ class TestLangfuseTracer:
 
             LangfuseTracer(tracer=MockTracer(), name="Haystack", public=False)
             assert "tracing is disabled" in caplog.text
-
-
-class TestSpanContext:
-    def test_post_init(self):
-        with pytest.raises(ValueError):
-            SpanContext(name=None, operation_name="operation_name", component_type=None, tags={}, parent_span=None)
-        with pytest.raises(ValueError):
-            SpanContext(name="name", operation_name=None, component_type=None, tags={}, parent_span=None)
-        with pytest.raises(ValueError):
-            SpanContext(
-                name="name",
-                operation_name="operation_name",
-                component_type=None,
-                tags={},
-                parent_span=None,
-                trace_name=None,
-            )
-
-
-class TestDefaultSpanHandler:
-    def test_handle_generator(self):
-        mock_span = Mock()
-        mock_span.raw_span.return_value = mock_span
-        mock_span.get_data.return_value = {
-            "haystack.component.type": "OpenAIGenerator",
-            "haystack.component.output": {"replies": ["This the LLM's response"], "meta": [{"model": "test_model"}]},
-        }
-
-        handler = DefaultSpanHandler()
-        handler.handle(mock_span, component_type="OpenAIGenerator")
-
-        assert mock_span.update.call_count == 1
-        assert mock_span.update.call_args_list[0][1] == {"usage": None, "model": "test_model"}
-
-    def test_handle_chat_generator(self):
-        mock_span = Mock()
-        mock_span.raw_span.return_value = mock_span
-        mock_span.get_data.return_value = {
-            "haystack.component.type": "OpenAIChatGenerator",
-            "haystack.component.output": {
-                "replies": [
-                    ChatMessage.from_assistant(
-                        "This the LLM's response",
-                        meta={"model": "test_model", "completion_start_time": "2021-07-27T16:02:08.012345"},
-                    )
-                ]
-            },
-        }
-
-        handler = DefaultSpanHandler()
-        handler.handle(mock_span, component_type="OpenAIChatGenerator")
-
-        assert mock_span.update.call_count == 1
-        assert mock_span.update.call_args_list[0][1] == {
-            "usage": None,
-            "model": "test_model",
-            "completion_start_time": datetime.datetime(2021, 7, 27, 16, 2, 8, 12345),
-        }
-
-    def test_handle_bad_completion_start_time(self, caplog):
-        mock_span = Mock()
-        mock_span.raw_span.return_value = mock_span
-        mock_span.get_data.return_value = {
-            "haystack.component.type": "OpenAIChatGenerator",
-            "haystack.component.output": {
-                "replies": [
-                    ChatMessage.from_assistant(
-                        "This the LLM's response",
-                        meta={"model": "test_model", "completion_start_time": "2021-07-32"},
-                    )
-                ]
-            },
-        }
-
-        handler = DefaultSpanHandler()
-        with caplog.at_level(logging.ERROR):
-            handler.handle(mock_span, component_type="OpenAIChatGenerator")
-            assert "Failed to parse completion_start_time" in caplog.text
-
-        assert mock_span.update.call_count == 1
-        assert mock_span.update.call_args_list[0][1] == {
-            "usage": None,
-            "model": "test_model",
-            "completion_start_time": None,
-        }
